Tests/LibWeb/Text/expected/css-has-invalidation/has-complex-sibling-subtree-scan-insights.txt

hasAncestorWalkInvocations: 1
hasAncestorWalkVisits: 5
hasAncestorSiblingElementChecks: 3
hasMatchInvocations: 1
hasResultCacheHits: 0
hasResultCacheMisses: 1
styleInvalidations: 1
